requirements.txt
bcrypt==4.1.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==2.2.3
WTForms==3.1.1
wtforms-validators==1.0.0
